o Update test for timeb.

+++ timeb-test.c
@@ -4,7 +4,7 @@
 #include <sys/timeb.h>
 #include <sys/time.h>
 
-long double
+double
 get_current_time_ms()
 {
         struct timeval tv;
@@ -20,6 +20,6 @@ main() {
         printf("time: %hu\n", ftm.millitm);
         ftime(&ftm);
         printf("time: %hu\n", ftm.millitm);
-        printf("%.3Lf\n", get_current_time_ms() / 1000);
+        printf("%f\n", get_current_time_ms() / 1000);
         return 0;
 }
